moved function out to separate file to clean up and centralize filter function
